+++ NEWS.md
@@ -1,5 +1,10 @@
 # purrr 0.2.2.9000
 
+* New `modify()` is shorthand for `x[] <- map(x, f)`.
+
+* `map_at()` and `map_if()` have been deprecated. Please use `modify_at()` 
+  and `modify_if()` instead (#292).
+
 * `invoke()` no longer uses lazyeval to figure out which enviroment a 
   character `f` comes from.